General cleanups

+++ src/buffer.c
@@ -10,7 +10,7 @@ void commitLine(tsState *psState) {
     allocLine(psState, psState->lineY, psState->editBuffer);
 }
 
-void loadLine(tsState *psState, int lineIndex) {
+void loadLine(tsState *psState, uint16_t lineIndex) {
     if (psState->text[lineIndex]) {
         strncpy(psState->editBuffer, psState->text[lineIndex], MAX_LINE_LENGTH - 1);
         psState->editBuffer[MAX_LINE_LENGTH - 1] = '\0';
